Simplify ReaderMenuStore
